requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
pyinstaller>=6.0.0
